refactor(common): fix a typo in package naming
