+++ answer-service/src/main/java/lk/ac/sjp/foe/co4353/g6/answerservice/AnswerServiceApplication.java
@@ -1,0 +1,15 @@
+package lk.ac.sjp.foe.co4353.g6.answerservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class AnswerServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AnswerServiceApplication.class, args);
+    }
+
+}
