+++ discovery-service/src/main/java/lk/ac/sjp/foe/co4353/g6/discoveryservice/DiscoveryServiceApplication.java
@@ -1,0 +1,15 @@
+package lk.ac.sjp.foe.co4353.g6.discoveryservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@SpringBootApplication
+@EnableEurekaServer
+public class DiscoveryServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DiscoveryServiceApplication.class, args);
+    }
+
+}
