--- discovery-service/src/test/java/lk/ac/sjp/foe/co4353/g5/discoveryservice/DiscoveryServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package lk.ac.sjp.foe.co4353.g5.discoveryservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DiscoveryServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
