+++ discovery-service/src/test/java/lk/ac/sjp/foe/co4353/g6/discoveryservice/DiscoveryServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package lk.ac.sjp.foe.co4353.g6.discoveryservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DiscoveryServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
