+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>lk.ac.sjp.foe.co4353.g5</groupId>
+    <groupId>lk.ac.sjp.foe.co4353.g6</groupId>
     <artifactId>microservices-demo</artifactId>
     <version>1.0-SNAPSHOT</version>
     <name>microservices-demo</name>
