+++ question-service/pom.xml
@@ -8,7 +8,7 @@
         <version>2.4.5</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
-    <groupId>lk.ac.sjp.foe.co4353.g5</groupId>
+    <groupId>lk.ac.sjp.foe.co4353.g6</groupId>
     <artifactId>question-service</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>question-service</name>
