+++ question-service/src/main/java/lk/ac/sjp/foe/co4353/g6/questionservice/QuestionServiceApplication.java
@@ -1,0 +1,15 @@
+package lk.ac.sjp.foe.co4353.g6.questionservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class QuestionServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(QuestionServiceApplication.class, args);
+    }
+
+}
