--- question-service/src/test/java/lk/ac/sjp/foe/co4353/g5/questionservice/QuestionServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package lk.ac.sjp.foe.co4353.g5.questionservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class QuestionServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
