+++ question-service/src/test/java/lk/ac/sjp/foe/co4353/g6/questionservice/QuestionServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package lk.ac.sjp.foe.co4353.g6.questionservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class QuestionServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
