--- reputation-service/src/test/java/lk/ac/sjp/foe/co4353/g5/reputationservice/ReputationServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package lk.ac.sjp.foe.co4353.g5.reputationservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ReputationServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
