+++ user-service/src/main/java/lk/ac/sjp/foe/co4353/g6/userservice/UserServiceApplication.java
@@ -1,0 +1,15 @@
+package lk.ac.sjp.foe.co4353.g6.userservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class UserServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(UserServiceApplication.class, args);
+    }
+
+}
