+++ user-service/src/test/java/lk/ac/sjp/foe/co4353/g6/userservice/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package lk.ac.sjp.foe.co4353.g6.userservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
