--- vote-service/src/main/java/lk/ac/sjp/foe/co4353/g5/voteservice/VoteServiceApplication.java
@@ -1,15 +1,0 @@
-package lk.ac.sjp.foe.co4353.g5.voteservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-@SpringBootApplication
-@EnableEurekaClient
-public class VoteServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(VoteServiceApplication.class, args);
-    }
-
-}
