--- vote-service/src/test/java/lk/ac/sjp/foe/co4353/g5/voteservice/VoteServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package lk.ac.sjp.foe.co4353.g5.voteservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class VoteServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
