+++ vote-service/src/test/java/lk/ac/sjp/foe/co4353/g6/voteservice/VoteServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package lk.ac.sjp.foe.co4353.g6.voteservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class VoteServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
