--- zuul-gateway/src/test/java/lk/ac/sjp/foe/co4353/g5/zuulgateway/ZuulGatewayApplicationTests.java
@@ -1,17 +1,0 @@
-package lk.ac.sjp.foe.co4353.g5.zuulgateway;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
-
-@SpringBootTest
-@EnableZuulProxy
-@EnableDiscoveryClient
-class ZuulGatewayApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
